refactor: extract acme plugin models into separate file
